add roles

+++ internal/service/user_service.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/scharph/orda/internal/domain"
 	"github.com/scharph/orda/internal/ports"
@@ -9,13 +10,14 @@ import (
 )
 
 type UserService struct {
-	repo ports.IUserRepository
+	repo  ports.IUserRepository
+	roles ports.IRoleRepository
 }
 
 var _ ports.IUserService = (*UserService)(nil)
 
-func NewUserService(repo ports.IUserRepository) *UserService {
-	return &UserService{repo}
+func NewUserService(repo ports.IUserRepository, roles ports.IRoleRepository) *UserService {
+	return &UserService{repo, roles}
 }
 
 func (s *UserService) GetAllUsers(ctx context.Context) ([]ports.UserResponse, error) {
@@ -34,41 +36,52 @@ func (s *UserService) GetAllUsers(ctx context.Context) ([]ports.UserResponse, er
 	return users, nil
 }
 
-func (s *UserService) CreateUser(ctx context.Context, req ports.UserRequest) (ports.UserResponse, error) {
+func (s *UserService) Create(ctx context.Context, req ports.UserRequest) (*ports.UserResponse, error) {
+	if req.Username == "admin" || req.Role == "admin" {
+		return nil, fmt.Errorf("admin can not be created")
+	} else if req.Username == "" || req.Password == "" || req.Role == "" {
+		return nil, fmt.Errorf("invalid input")
+	}
+
+	_, err := s.roles.ReadByName(ctx, req.Role)
+	if err != nil {
+		return nil, fmt.Errorf("role not found")
+	}
+
 	pw, err := util.HashPassword(req.Password)
 	if err != nil {
-		return ports.UserResponse{}, err
+		return nil, err
 	}
 
 	user, err := s.repo.Create(ctx, &domain.User{Username: req.Username, Password: pw, Role: req.Role})
 	if err != nil {
-		return ports.UserResponse{}, err
+		return nil, err
 	}
-	return ports.UserResponse{
+	return &ports.UserResponse{
 		Username: user.Username,
 		Role:     user.Role,
 		Id:       user.ID,
 	}, nil
 }
 
-func (s *UserService) GetUserById(ctx context.Context, id string) (ports.UserResponse, error) {
+func (s *UserService) GetUserById(ctx context.Context, id string) (*ports.UserResponse, error) {
 	user, err := s.repo.ReadById(ctx, id)
 	if err != nil {
-		return ports.UserResponse{}, err
+		return nil, err
 	}
-	return ports.UserResponse{
+	return &ports.UserResponse{
 		Username: user.Username,
 		Role:     user.Role,
 		Id:       user.ID,
 	}, nil
 }
 
-func (s *UserService) GetUserByUsername(ctx context.Context, username string) (ports.UserResponse, error) {
+func (s *UserService) GetUserByUsername(ctx context.Context, username string) (*ports.UserResponse, error) {
 	user, err := s.repo.ReadByUsername(ctx, username)
 	if err != nil {
-		return ports.UserResponse{}, err
+		return nil, err
 	}
-	return ports.UserResponse{
+	return &ports.UserResponse{
 		Username: user.Username,
 		Role:     user.Role,
 		Id:       user.ID,
@@ -76,11 +89,20 @@ func (s *UserService) GetUserByUsername(ctx context.Context, username string) (p
 	}, nil
 }
 
-func (s *UserService) UpdateUser(ctx context.Context, req ports.UserRequest) (ports.UserResponse, error) {
-	return ports.UserResponse{}, nil
+func (s *UserService) Update(ctx context.Context, req ports.UserRequest) (*ports.UserResponse, error) {
+	current, _ := s.GetUserById(ctx, req.Id)
+	// Prevent username "admin" to be updated
+	if current.Username == "admin" && req.Username != "admin" {
+		return nil, fmt.Errorf("admin can not be updated")
+	}
+	return nil, fmt.Errorf("Not implemented")
 }
 
-func (s *UserService) DeleteUser(ctx context.Context, id string) error {
+func (s *UserService) Delete(ctx context.Context, id string) error {
+	u, _ := s.GetUserById(ctx, id)
+	if u.Username == "admin" {
+		return fmt.Errorf("admin can not be deleted")
+	}
 	_, err := s.repo.Delete(ctx, id)
 	return err
 }
